pyproject.toml
[tool.poetry]
name = "cadenai"
version = "0.2.0"
description = ""
authors = ["Savombre <clementrobin6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
openai = "^1.6.1"
pydantic = "^2.5.3"
pypdf = "^3.17.4"
qdrant-client = "^1.7.0"
tenacity = "^8.2.3"
tiktoken = "^0.5.2"
mistralai = "^0.0.9"

[tool.poetry.group.dev.dependencies]
jupyter-client = "^8.6.0"
jupyter-core = "^5.6.1"
pipreqs = "^0.4.13"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
